change tasks from file to dir
