+++ hetseq/tasks/__init__.py
@@ -1,0 +1,11 @@
+from .tasks import Task, LanguageModelingTask, MNISTTask
+from .bert_for_el_classification_task import BertForELClassificationTask
+from .bert_for_token_classification_task import BertForTokenClassificationTask
+
+__all__ = [
+    'Task',
+    'LanguageModelingTask',
+    'BertForTokenClassificationTask',
+    'MNISTTask',
+    'BertForELClassificationTask',
+]
